Changes to "duplicate observation"
This change is for code on lines 63-75, where the exclusion note "duplicate observation" is made for some observations.

We updated the Admin database (date 11-12) to fix issues with some of these duplicates. For Cuba 2003 we deleted duplicates, for Germany 2014, the duplicate was due to changing not defined. We should not change not defined to 28wks for this since there is a 28wk observation. The definition_rv is changed back to "not defined". We think the Belgium 2004 definition is no longer duplicate since the 1000g and 28wk  is not being changed to 1000g.