Use HQ LMIC data for definition adjustment
For denominational adjustment should only use high quality data for LMIC (so not admin data from the full dataset). I made the change here and tried to ensure the output saved is the same. I didn't apply your function find_comb_def_adj( ), because the dataset was already pretty close to the output of your function. 

*** IMPORTANT Question:  Is this the only place the change needs to be made? I couldn't tell in the version of 3.def_adj_reg where the def_adj_list_lmic output is used, so I wasn't sure if more changes need to be made.